+++ include/rtmp.h
@@ -7799,7 +7799,7 @@ VOID MaintainBssTable(
 	IN  UCHAR	MaxSameRxTimeCount);
 #endif /* CONFIG_STA_SUPPORT */
 
-void  getRate(HTTRANSMIT_SETTING HTSetting, ULONG* fLastTxRxRate);
+void  getRate(HTTRANSMIT_SETTING HTSetting, uint32_t *fLastTxRxRate);
 
 
 #ifdef APCLI_SUPPORT
